CharacterAiChatBot/requirements.txt
langchain
flask
flask-cors
openai==0.27.8
google-analytics-python